mcp-servers/dynamodb/requirements.txt
# MCP Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Environment & Configuration
python-dotenv==1.0.0

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
hypothesis==6.92.1
